requirements.txt
pytesseract
Pillow
pdfplumber
transformers
torch
datasets
fastapi
uvicorn
